[LIB-795] Move rename() method from model instances to shared module
