+++ src/models/dataendpoint.js
@@ -1,6 +1,6 @@
 import stampit from 'stampit';
 import _ from 'lodash';
-import {Meta, Model} from './base';
+import {Meta, Model, Rename} from './base';
 import QuerySet, {Filter} from '../querySet';
 
 
@@ -157,6 +157,7 @@ const DataEndpoint = stampit()
   .setMeta(DataEndpointMeta)
   .setQuerySet(DataEndpointQerySet)
   .setConstraints(DataEndpointConstraints)
+  .compose(Rename)
   .methods({
 
     /**
@@ -180,26 +181,6 @@ const DataEndpoint = stampit()
       if(!_.isEmpty(filters)) query.query = JSON.stringify(filters)
 
       return this.makeRequest('GET', path, {query});
-    },
-    /**
-    * Renames DataEndpoint.
-    * @memberOf DataEndpoint
-    * @instance
-
-    * @param {Object} payload object containing the payload to be sent
-    * @returns {Promise}
-
-    * @example {@lang javascript}
-    * DataEndpoint.rename({ new_name: 'new_name'}).then(function(dataendpoint) {});
-    */
-    rename(payload = { new_name: this.name }) {
-      const meta = this.getMeta();
-      const path = meta.resolveEndpointPath('rename', this);
-
-      return this.makeRequest('POST', path, {payload})
-        .then((response) => {
-          return this.serialize(response);
-        })
     },
     /**
     * Clears cache of DataEndpoint.
